Initial commit
This program crawls Github repository information automatically.
The program requires two arguments when running. The first argument is
the local path to a Github project. The second argument is a path to a
database file generated by Tacoco.

GitLog gets SLOC, the list of Authors, the history of authors regard to
each source code line, Commit information, the list of Co-committed
files, and the history of file name change.

GitBlame crawls the ‘git blame’ result and store the result into a
database.

PathProcessing finds the local path of each class file and stores the
result into a database.